feat(game): handle disconnection 로직 추가

+++ backend/src/game/game.gateway.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/websockets';
 import { Socket, Server } from 'socket.io';
 import { GameService } from './game.service';
-import { BallInfo, PlayerNumber, Score } from 'src/type/game';
+import { BallInfo, PlayerNumber, Score, GameInfo } from 'src/type/game';
 import { parse } from 'cookie';
 import { JwtService } from '@nestjs/jwt';
 import { UseGuards } from '@nestjs/common';
@@ -27,26 +27,37 @@ export class GameGateway {
   @WebSocketServer() server: Server;
   fps = 1000 / 60;
 
-  async handleConnection(client: Socket) {
+  async handleConnection(@ConnectedSocket() client: Socket) {
     console.log('handleConnection', client.id);
   }
 
-  async handleDisconnect(client: Socket) {
-    console.log('handleDisconnect', client.id);
-    this.gameService.removeMatchingClient(client);
+  async handleDisconnect(@ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
+    const roomName = this.gameService.getGameRoom(client);
+    if (!roomName) {
+      console.log('INFO handleDisconnect : 게임 방이 없습니다.');
+      this.gameService.removeMatchingClient(client);
+      return ;
+    }
+    const gameInfo = this.gameService.getGameInfo(roomName);
+    if (!gameInfo) {
+      console.log('INFO handleDisconnect : 게임 정보가 없습니다.');
+      client.leave(roomName);
+      return ;
+    }
+    this.server.to(roomName).emit('game-disconnect', { disconnectNickname: payload.nickname, gameInfo } );
   }
 
+
   @SubscribeMessage('game-randomStart')
-  handleStartGame(@MessageBody() data: any, @ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
-    const nickname = payload.intraNickname;
-    const [ roomName, player1Id, player2Id ] = this.gameService.match(client, nickname);
+  handleStartGame(@ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
+    const nickname = payload.nickname; const [ roomName, player1Id, player2Id ] = this.gameService.match(client, nickname);
     if (!roomName) this.server.to(client.id).emit('game-loading');
     if (!player1Id || !player2Id) return;
     this.server.to(roomName).emit('game-randomStart', {player1Id, player2Id});
   }
 
   @SubscribeMessage('game-keyEvent')
-  handleGameKeyEvent(@MessageBody() data: any, @ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
+  handleGameKeyEvent(@ConnectedSocket() client: Socket, @MessageBody() data: any, @PnJwtPayload() payload: PnPayloadDto) {
     const roomName = this.gameService.getGameRoom(client);
     console.log('roomName', roomName);
     this.server.to(data.opponentId).emit('game-keyEvent', {
@@ -59,8 +70,7 @@ export class GameGateway {
 
   // TODO: sensor에 닿을 시 score 변경
   @SubscribeMessage('game-score')
-  handleScore(@MessageBody() data: {playerNumber: PlayerNumber, score: Score}, @ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
-    console.log('INFO: game-score', data);
+  handleScore(@ConnectedSocket() client: Socket, @MessageBody() data: {playerNumber: PlayerNumber, score: Score}, @PnJwtPayload() payload: PnPayloadDto) {
     const roomName = this.gameService.getGameRoom(client);
     const gameInfo = this.gameService.getGameInfo(roomName);
     if (!gameInfo) return ;
@@ -75,7 +85,7 @@ export class GameGateway {
   }
 
   @SubscribeMessage('game-ball')
-  handleBall(@MessageBody() ball: BallInfo, @ConnectedSocket() client: Socket, @PnJwtPayload() payload: PnPayloadDto) {
+  handleBall(@ConnectedSocket() client: Socket, @MessageBody() ball: BallInfo) {
     const roomName = this.gameService.getGameRoom(client);
     const updatedBallInfo = this.gameService.reconcilateBallInfo(roomName, ball);
     if (!updatedBallInfo) return;
